refactor(runner): one subsystem, one module — main.py 1213 -> 596 lines
Pure movement, no behaviour change. `main.py` had grown into a god-module: 28% of
the package, 34 top-level functions, seven distinct responsibilities. The specific
smell was not size but that for most of them a module ALREADY EXISTED holding only
half the job — menu.py parsed dialogs while main answered them; chat_bridge.py held
the bridge while main pumped it; inbox.py/schedules.py had their triggers in main.
Editing safely meant grepping 1200 lines to work out which subsystem you were in.

Now every subsystem owns its module, and main.py is the loop and the CLI:

  sessions.py     reporting the open emdash sessions this box can see
  chat_pump.py    driving a chat turn to completion across ticks
  streams.py      live transcript push while a viewer is attached
  hooks.py        hook tool-events, and answering a blocked agent's dialog
  cancel.py       the shared "stop this turn" set
  failure_log.py  repeat-aware logging for the best-effort retry paths

Two structural rules, both enforced by the import graph rather than by comment:

- **Dependencies point one way.** main imports its subsystems, never the reverse.
  CANCELLED_TURNS and the retry logger became leaf modules precisely because they
  are shared BY chat_pump and streams and PRODUCED by the loop — parking either in
  main would have made those modules import their own caller.
- **Menu parsing and menu answering are different jobs.** Parsing stays in menu.py
  (pure, screen in / dict out); answering moved to hooks.py, because it resolves a
  cwd/session_key to an emdash task and is hook-shaped. Splitting them the other
  way makes menu.py and hooks.py import each other.

Also moves tree.py + render_tree_svg.py OUT of the package to assets/brand/, beside
their only consumer. They are icon geometry the daemon never imports, and living in
the package meant 112 lines of it were built into the wheel and installed onto every
box in the fleet. A packaging test now asserts they stay out.

Cross-module callees became public (module.foo, not module._foo); names still
internal to a module kept their underscore. Nine test files turned out to no longer
import main at all once rewritten — ruff flagged the now-unused imports, which is
the decoupling made visible.

Verified beyond the suites: built and installed the refactored package into an
isolated tool dir and captured its heartbeat on the wire — byte-identical payload
to before, same session report, same CDP preflight. 289 runner tests, 1900 backend
(including the cross-seam menu round-trip, which had to be repointed at hooks),
76 cloud_runner, frontend vitest, ruff.
