+++ packages/canopy_runner/canopy_runner/hooks.py
@@ -1,0 +1,239 @@
+"""The hook-driven half of a live session: tool events, and the dialog a blocked
+agent is waiting on.
+
+Claude Code fires a hook per tool call straight to a loopback listener this
+runner owns — the LIVE half of a session's record (the transcript is complete
+but lags, so it cannot drive a view you are actively watching).
+
+Menu ANSWERING lives here rather than in `menu.py` because it is hook-shaped:
+it resolves a cwd/session_key to an emdash task. `menu.py` stays the pure
+parser, which is what keeps this dependency one-directional."""
+from __future__ import annotations
+
+import logging
+import time
+import uuid
+from pathlib import Path
+
+import canopy_transcript as transcript_core
+
+from . import cdp_control, hook_install, menu
+from .client import Client
+from .config import Config
+
+logger = logging.getLogger("canopy_runner")
+
+
+# --- Live hook events (spec 2026-07-27) -------------------------------------
+#
+# Claude Code fires a hook per tool call straight to a loopback listener this
+# runner owns. That is the LIVE half of a session's record: the transcript is
+# complete but lags (the docs say so explicitly), so it cannot drive a view you
+# are actively watching. The transcript remains the durable record, which is
+# what makes it safe for this path to drop events freely.
+#
+# `{(project, session_key): session_id}`, refreshed from the stream sync each
+# tick — the hook reports a cwd, and this is what turns that back into a canopy
+# Session.
+_hook_sessions: dict[tuple[str, str], str] = {}
+_hook_listener = None
+
+
+# None, NOT 0.0: `time.monotonic()` is near zero early in a process's life, so
+# seeding this with 0.0 reads as "already reported just now" and suppresses the
+# first report for a whole window — exactly when you are watching for it, because
+# you have just turned the feature on. Found on the first live enablement; the
+# original test hid it by starting its fake clock at 10,000.
+_last_hook_report: float | None = None
+# Same cadence as the idle-cycle line: often enough to answer "is it working?"
+# without turning a busy agent into a log flood.
+HOOK_REPORT_SECONDS = 300
+
+
+def maybe_report_hooks(now_fn=time.monotonic) -> None:
+    """Log what the hook listener has seen.
+
+    Without this the live path is unobservable from the runner side: events are
+    accepted, resolved and forwarded entirely silently, so "is it working?" can
+    only be answered from the server's logs — which is exactly the question you
+    have when you cannot reach them. Counters are cumulative since start.
+    """
+    global _last_hook_report
+    listener = _hook_listener
+    if listener is None:
+        return
+    if _last_hook_report is not None and now_fn() - _last_hook_report < HOOK_REPORT_SECONDS:
+        return
+    if listener.received == 0:
+        # Nothing has fired yet; silence is the honest report — and crucially we
+        # do NOT stamp, or an idle tick right after startup burns the whole
+        # window and the first real report waits 5 minutes behind it.
+        return
+    _last_hook_report = now_fn()
+    logger.info(
+        "hooks: %d received, %d forwarded, %d dropped (cwd not a session we back), "
+        "forwarding=%s",
+        listener.received, listener.forwarded, listener.dropped_unknown_cwd,
+        listener._forward(),
+    )
+
+
+def resolve_hook_session(cwd: str) -> str:
+    """A hook's cwd -> canopy session id, or "" if this isn't a session we back.
+
+    Hooks are installed at USER level, so they fire for every Claude Code
+    session on the machine. Most are not ours; returning "" is the expected
+    path, not a failure.
+    """
+    if not cwd:
+        return ""
+    parsed = transcript_core.parse_emdash_worktree(cwd, home=Path.home())
+    if parsed is None:
+        return ""
+    project, task = parsed
+    # The worktree dir may carry emdash's random de-dupe suffix, so try the exact
+    # name first and the stripped one second — matching against sessions we
+    # actually know rather than guessing which it is.
+    for candidate in transcript_core.emdash_task_candidates(task):
+        session_id = _hook_sessions.get((project, candidate))
+        if session_id:
+            return session_id
+    return ""
+
+
+def hook_task_name(cwd: str) -> str:
+    """The emdash task backing this cwd, or "" — the handle CDP drives by."""
+    if not cwd:
+        return ""
+    parsed = transcript_core.parse_emdash_worktree(cwd, home=Path.home())
+    if parsed is None:
+        return ""
+    project, task = parsed
+    for candidate in transcript_core.emdash_task_candidates(task):
+        if _hook_sessions.get((project, candidate)):
+            return candidate
+    return ""
+
+
+# The two functions below take the CDP module as an argument so the whole
+# chain — screen in, keystrokes out — is testable against a captured terminal
+# with no emdash, no browser and no Playwright. The `cdp_control`-bound wrappers
+# under them are what the runner actually calls.
+
+
+
+def read_hook_menu_from(cdp, task: str, *, cdp_port: int = 9222):
+    """The dialog on `task`'s screen as a plain dict, or None."""
+    if not task:
+        return None
+    found = menu.find_menu(cdp.read_terminal(task, port=cdp_port))
+    if found is None:
+        return None
+    return {
+        "question": found.question,
+        "title": found.title,
+        "body": found.body,
+        "selected": found.selected,
+        "options": [{"number": o.number, "label": o.label} for o in found.options],
+    }
+
+
+def answer_menu_with(cdp, session_key: str, option, *, cdp_port: int = 9222) -> None:
+    """Press a human's answer into `session_key`'s terminal.
+
+    Re-reads the screen FIRST and refuses an option that is not on it. A menu can
+    go stale between the phone rendering it and a thumb reaching it — and a
+    NUMBER typed at a session no longer showing a dialog lands in its prompt,
+    where the agent reads a bare "1" as an instruction. Double-taps and two
+    people answering at once both land here.
+    """
+    # Re-read with a settle: a single read can catch the TUI mid-render, and
+    # dropping a human's tap because the footer had not painted yet is a bug
+    # they experience as "the button did nothing".
+    current = menu.find_menu_settled(
+        lambda: cdp.read_terminal(session_key, port=cdp_port))
+    if current is None:
+        logger.info("menu answer for %s ignored — no dialog on screen now", session_key)
+        return
+    number = None if option is None else int(option)
+    if not current.allows(number):
+        logger.warning("menu answer %r for %s is not on the dialog now showing (%d options)",
+                       number, session_key, len(current.options))
+        return
+    cdp.send_keys(session_key, menu.answer_keys(number), port=cdp_port)
+    logger.info("answered the dialog on %s with %s", session_key,
+                "Esc" if number is None else f"option {number}")
+
+
+def read_hook_menu(cwd: str, *, cdp_port: int):
+    """The dialog on this session's screen, or None.
+
+    Only called when a session reports BLOCKED: a hook can say an agent wants a
+    human but never what it is asking, and emdash owns the session, so the
+    question and its options exist only on the terminal.
+    """
+    return read_hook_menu_from(cdp_control, hook_task_name(cwd), cdp_port=cdp_port)
+
+
+def answer_menu(session_key: str, option, *, cdp_port: int = 9222) -> None:
+    answer_menu_with(cdp_control, session_key, option, cdp_port=cdp_port)
+
+
+
+def start_hook_listener(cfg: Config, client: Client):
+    """Install the user-level hook and start the loopback listener.
+
+    Returns the listener, or None when disabled (`hook_port = 0`), in which case
+    any previously-installed canopy hook is REMOVED — turning the feature off
+    must not leave a curl pointing at a port nothing is listening on.
+    """
+    global _hook_listener
+    settings_path = Path.home() / ".claude" / "settings.json"
+    if cfg.hook_port <= 0:
+        if hook_install.remove(settings_path):
+            logger.info("hook listener disabled; removed canopy's hook from %s",
+                        settings_path)
+        return None
+    from .hook_listener import HookListener
+
+    nonce = uuid.uuid4().hex
+    # NO read_menu here, deliberately. Reading a session's screen means driving
+    # emdash over CDP, and `openTask` CLICKS the task in the sidebar and focuses
+    # its terminal — so wiring it to a hook meant every Notification yanked
+    # emdash to whatever agent had just asked for input, mid-typing.
+    #
+    # Reported 2026-07-28, twice: focus taken, a few characters typed into the
+    # newly-focused prompt, then the message that was meant for that task never
+    # arrived. The second half is the collision guard working exactly as designed
+    # — `open_and_send` found unsent text, refused to clobber it, and defaulted to
+    # a fresh session — so this bug MANUFACTURED the leaked-keystroke case that
+    # guard exists to catch.
+    #
+    # A menu can still be read on demand (`cdp_control.read_terminal`), where the
+    # task switch is something a human just asked for. It cannot be read on a
+    # signal, because there is no way to read a NON-active task: emdash marks no
+    # task as current in the DOM (checked — no aria-current, no data-state), so
+    # identifying whose screen you are looking at requires activating it first.
+    listener = HookListener(
+        port=cfg.hook_port, nonce=nonce,
+        resolve_session=resolve_hook_session,
+        forward=lambda: cfg.forward_sessions,
+    )
+    listener.bind_sender(
+        lambda session_id, events: client.post_session_stream(
+            cfg.runner_id, session_id, events)
+    )
+    try:
+        listener.start()
+    except OSError as exc:
+        # Another process already holds the port (a second runner, a stale
+        # process). Live events are an overlay, so this is a warning, not fatal.
+        logger.warning("hook listener could not bind :%d (%s); live events off",
+                       cfg.hook_port, exc)
+        return None
+    hook_install.install(settings_path, port=cfg.hook_port, nonce=nonce)
+    logger.info("live hook events: listener on :%d, forwarding=%s",
+                cfg.hook_port, cfg.forward_sessions)
+    _hook_listener = listener
+    return listener
+
