+++ packages/canopy_runner/tests/test_cdp_control.py
@@ -262,51 +262,3 @@ def test_ensure_sidecar_deps_fails_when_npm_claims_success_but_installs_nothing(
                         lambda *a, **k: SimpleNamespace(stdout="", stderr="", returncode=0))
     with pytest.raises(cdp_control.CDPError):
         cdp_control.ensure_sidecar_deps()
-
-
-# -- the evaluated JS must be valid, which node --check cannot tell us ---------
-
-def _evaluate_bodies():
-    """Every string this sidecar hands to page.evaluate, as Playwright sees it."""
-    import re
-    from pathlib import Path
-    src = (Path(__file__).resolve().parents[1] / "canopy_runner" / "cdp"
-           / "emdash_control.mjs").read_text()
-    helper = re.search(r"const ACTIVE_TERM_FN = String\.raw`([\s\S]*?)`;", src)
-    helper_src = helper.group(1) if helper else ""
-    return [m.replace("${ACTIVE_TERM_FN}", helper_src)
-            for m in re.findall(r"page\.evaluate\(String\.raw`([\s\S]*?)`\)", src)]
-
-
-def test_every_evaluate_string_is_valid_javascript():
-    """`node --check` passes on a file whose TEMPLATE contents are broken, because
-    the template is a valid string literal either way. What Playwright receives is
-    the PROCESSED text — and a plain template ate the escapes: `\\n` became a real
-    newline (unterminated string -> "SyntaxError: Invalid or unexpected token" at
-    runtime) and `\\s` collapsed to a literal `s`, turning every `\\s*` into "zero
-    or more letter s". The collision guard silently stopped matching and nothing
-    in CI noticed. String.raw is what prevents it; this test is what catches a
-    regression.
-    """
-    import shutil
-    import subprocess
-    node = shutil.which("node")
-    if node is None:                       # CI without node: nothing to assert
-        return
-    bodies = _evaluate_bodies()
-    assert bodies, "no page.evaluate(String.raw`…`) sites found — did the shape change?"
-    for i, body in enumerate(bodies, start=1):
-        check = subprocess.run(
-            [node, "-e", "new Function('return ' + require('fs').readFileSync(0, 'utf8'))"],
-            input=body, capture_output=True, text=True, timeout=30)
-        assert check.returncode == 0, f"evaluate #{i} is invalid JS: {check.stderr[:300]}"
-
-
-def test_the_evaluate_sites_use_string_raw():
-    """A plain backtick here reintroduces the escape-eating bug."""
-    import re
-    from pathlib import Path
-    src = (Path(__file__).resolve().parents[1] / "canopy_runner" / "cdp"
-           / "emdash_control.mjs").read_text()
-    assert "page.evaluate(`" not in src, "use String.raw` for evaluate source"
-    assert len(re.findall(r"page\.evaluate\(String\.raw`", src)) >= 4
